Fixes a few minor problems, adds a lot of documentation, add a new command.
The new command is as follows:

Search a given file for a pattern:
 /some/file:/REGEX

Search a dir for a pattern:
 /some/dir/*:/REGEX

Recursively search a dir for a pattern:
 /some/dir/**:/REGEX

Search uses grep -E patterns, binary files are not searched.
If there are multiple matches, dmenu is used to let the
user select the correct choice, then open the file with editor.